Support paused containers when pre-heating.
Prior to this change, allcontainers will emit start for existing
"preheat" containers even if they are paused, but you cannot
attach to a paused container with docker, so consumers of this
event end up getting an error (see e.g. docker-loghose).

This change will fire the start event if/when the existing
container becomes unpaused (only once), preventing these errors.